NF: move formatting filters to template
